+++ ERGOCONCEPT/Assets/Scripts/UI/AccessorySeries.cs
@@ -29,7 +29,7 @@ public class AccessorySeries : Accessory
         GetComponent<RectTransform>().sizeDelta = new Vector2(currentSize.x, currentSize.y + rectSize);
     }
 
-    public override void ChangeValue(Slider slider, string parameterTypes, int multiplicateur, string axe)
+    public override void ChangeValue(Slider slider, string parameterTypes, float multiplicateur, string axe)
     {
         base.ChangeValue(slider, parameterTypes, multiplicateur, axe);
     }
